+++ src/bootstrap/check_config.ts
@@ -2,13 +2,21 @@ import { LooseObject } from '@balm/index';
 
 const docsBaseURL = 'https://balmjs.com/docs/v2/config';
 
+function upgradeGuide(key: string, value: string, message: string) {
+  return `${message} See ${docsBaseURL}/${key}.html#${key}-${value}`;
+}
+
 // Compatibility for upgrade balm
 function checkConfig(): void {
   // v2.22.0
   if ((BalmJS.config.images as LooseObject).defaultPlugins) {
     BalmJS.logger.warn(
       'balm@2.22.0+ config',
-      `'images.defaultPlugins' was renamed to 'images.plugins'. See ${docsBaseURL}/images.html#images-plugins`
+      upgradeGuide(
+        'images',
+        'plugins',
+        '`images.defaultPlugins` was renamed to `images.plugins`.'
+      )
     );
   }
 
@@ -16,7 +24,11 @@ function checkConfig(): void {
   if ((BalmJS.config.styles as LooseObject).postcssLoaderOptions) {
     BalmJS.logger.warn(
       'balm@2.11.0+ config',
-      `'styles.postcssLoaderOptions' was migrated to 'scripts.postcssLoaderOptions'. See ${docsBaseURL}/scripts.html#scripts-postcssloaderoptions`
+      upgradeGuide(
+        'scripts',
+        'postcssLoaderOptions',
+        '`styles.postcssLoaderOptions` was migrated to `scripts.postcssLoaderOptions`.'
+      )
     );
   }
 
@@ -24,7 +36,11 @@ function checkConfig(): void {
   if ((BalmJS.config.scripts as LooseObject).disableDefaultLoaders) {
     BalmJS.logger.warn(
       'balm@2.5.0+ config',
-      `'scripts.disableDefaultLoaders' was renamed to 'scripts.defaultLoaders'. See ${docsBaseURL}/scripts.html#scripts-defaultloaders`
+      upgradeGuide(
+        'scripts',
+        'defaultLoaders',
+        '`scripts.disableDefaultLoaders` was renamed to `scripts.defaultLoaders`.'
+      )
     );
   }
 }