+++ src/tasks/private/image.ts
@@ -57,7 +57,7 @@ class ImageTask extends BalmJS.BalmTask {
             BalmJS.logger.error(`${this.name} task`, error.message);
           })
         )
-        .pipe($.imagemin(this.plugins))
+        .pipe($.if(this.plugins.length, $.imagemin(this.plugins)))
         .pipe(gulp.dest(BalmJS.file.absPath(this.output)));
     };